[EHN] moving because of license
